complete all comments and regular the codes

+++ Wxt.OnlineSuperMarket.Business.Services/SuperMarketService.cs
@@ -1,17 +1,24 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Wxt.OnlineSuperMarket.Data.Entities;
-using Wxt.OnlineSuperMarket.Data.Repositories;
-
-namespace Wxt.OnlineSuperMarket.Business.Services
+﻿namespace Wxt.OnlineSuperMarket.Business.Services
 {
+    using System;
+    using Wxt.OnlineSuperMarket.Data.Entities;
+    using Wxt.OnlineSuperMarket.Data.Repositories;
+
+    /// <summary>
+    /// Defines the <see cref="SuperMarketService" />
+    /// </summary>
     public class SuperMarketService
     {
         private readonly ISuperMarketRepository _superMarketRepository = new InMemorySuperMarketRepository();
 
+        /// <summary>
+        /// The AddProuct method checks input and adds a new product to the backend repository.
+        /// </summary>
+        /// <param name="name">New product's name<see cref="string"/></param>
+        /// <param name="price">New product's price<see cref="decimal"/></param>
+        /// <param name="description">New product's description<see cref="string"/></param>
+        /// <param name="category">New product's category<see cref="Category"/></param>
+        /// <returns>The whole information of new added product<see cref="string"/></returns>
         public string AddProuct(string name, decimal price, string description = null, Category category = 0)
         {
             if (string.IsNullOrWhiteSpace(name))
@@ -24,7 +31,7 @@ namespace Wxt.OnlineSuperMarket.Business.Services
             }
             var product = new Product()
             {
-                Name = name,
+                Name = name.Trim(),
                 Category = category,
                 Price = price,
                 Description = description
@@ -32,27 +39,53 @@ namespace Wxt.OnlineSuperMarket.Business.Services
             return _superMarketRepository.AddProduct(product).ToString();
         }
 
-        public bool RemoveProduct(int productId)
+        /// <summary>
+        /// The RemoveProduct method removes an existing product from backend repository.
+        /// </summary>
+        /// <param name="productId">The Id of the product to be removed<see cref="int"/></param>
+        public void RemoveProduct(int productId)
         {
-            return _superMarketRepository.RemoveProduct(productId);
+            if (productId <= 0)
+            {
+                throw new ArgumentOutOfRangeException("Product's Id is always bigger than 0.");
+            }
+            _superMarketRepository.RemoveProduct(productId);
         }
 
-        public bool IncreaseStockt(int productId, int count)
+        /// <summary>
+        /// The IncreaseStockt method increases the count of stock of one particular product to backend repository.
+        /// </summary>
+        /// <param name="productId">The Id of the product to be increased<see cref="int"/></param>
+        /// <param name="count">The count to be increased<see cref="int"/></param>
+        public void IncreaseStockt(int productId, int count)
         {
+            if (productId <= 0)
+            {
+                throw new ArgumentOutOfRangeException("Product's Id is always bigger than 0.");
+            }
             if (count <= 0)
             {
                 throw new ArgumentOutOfRangeException($"Cannot increase {count} product to stock.");
             }
-            return _superMarketRepository.IncreaseStock(productId, count);
+            _superMarketRepository.IncreaseStock(productId, count);
         }
 
-        public bool DecreaseStockt(int productId, int count)
+        /// <summary>
+        /// The DecreaseStockt method decreases the count of stock of one particular product from backend repository.
+        /// </summary>
+        /// <param name="productId">The Id of the product to be decreased<see cref="int"/></param>
+        /// <param name="count">The count to be decreased<see cref="int"/></param>
+        public void DecreaseStockt(int productId, int count)
         {
+            if (productId <= 0)
+            {
+                throw new ArgumentOutOfRangeException("Product's Id is always bigger than 0.");
+            }
             if (count <= 0)
             {
                 throw new ArgumentOutOfRangeException($"Cannot decrease {count} product from stock.");
             }
-            return _superMarketRepository.DecreaseStock(productId, count);
+            _superMarketRepository.DecreaseStock(productId, count);
         }
     }
 }
